adequação ao pep8

+++ bot.py
@@ -530,15 +530,20 @@ async def status(ctx, *, status):
 
 @bot.command()
 async def atualizar(ctx):
-    await ctx.message.delete()
-    os.chdir('..')
-    os.system('sudo rm -R Marselo-Bot')
-    os.system('git clone https://github.com/Eduardo-Barreto/Marselo-Bot.git')
-    os.system('cp /home/pi/Desktop/tokens.py /home/pi/Desktop/Marselo-Bot')
-    os.chdir('Marselo-Bot')
-    os.system('clear')
-    os.system('python3 bot.py')
-    quit()
+    if ctx.author.id == tokens.eduardo_id:
+        await ctx.message.delete()
+        os.chdir('..')
+        os.system('sudo rm -R Marselo-Bot')
+        os.system(
+            'git clone https://github.com/Eduardo-Barreto/Marselo-Bot.git'
+        )
+        os.system(
+            'cp /home/pi/Desktop/tokens.py /home/pi/Desktop/Marselo-Bot'
+        )
+        os.chdir('Marselo-Bot')
+        os.system('clear')
+        os.system('python3 bot.py')
+        quit()
 
 
 bot.run(tokens.discord)
